src/requirements.txt
pyodbc==5.0.1
pymysql==1.1.0
pg8000==1.30.3
snowflake-sqlalchemy==1.5.1
databricks-sql-connector==2.9.3
sqlalchemy-bigquery==1.9.0
google-cloud-bigquery-storage==2.24.0
colorama==0.4.6
PyYAML==6.0.1
Pyjeb==0.2.1
numpy==1.26.3
pandas==2.1.4
openpyxl==3.1.2
sqlalchemy==1.4.51
pyspark==3.5.1
pandas-gbq==0.23.0
pydata-google-auth==1.8.2
